+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add format_text example. Contributed by @mvasi90.
 - Fix some clippy lints.
 - Add Font::get_name().
+- Add image::RgbImage::convert().
 
 ## [1.1.7] - 2021-08-04
 - Fix invalid cast in TextBuffer::selection_position().